chore: add docs and organize exports
