+++ packages/client/src/v2-events/features/workqueues/EventOverview/EventOverview.tsx
@@ -94,7 +94,6 @@ function EventOverviewFull({
       'event.assignedTo': assignedTo,
       flags: eventIndex.flags
     }
-
   const { getEventTitle } = useEventTitle()
   const { title } = getEventTitle(eventConfiguration, eventWithDrafts)
