+++ packages/events/src/router/middleware/validate/utils.ts
@@ -10,14 +10,13 @@
  */
 
 import { TRPCError } from '@trpc/server'
-import _ from 'lodash'
 import {
   ActionUpdate,
   errorMessages,
   LocationType,
   ValidatorContext
 } from '@opencrvs/commons/events'
-import { getOrThrow } from '@opencrvs/commons'
+import { getOrThrow, flattenEntries } from '@opencrvs/commons'
 import { getTokenPayload } from '@opencrvs/commons/authentication'
 import { getLeafLocationIds } from '@events/storage/postgres/events/locations'
 
@@ -52,29 +51,6 @@ export function throwWhenNotEmpty(errors: unknown[]) {
       message: JSON.stringify(errors)
     })
   }
-}
-
-function isObject(value: unknown): value is Record<string, unknown> {
-  return _.isPlainObject(value)
-}
-
-/**
- * Flattens an object into a list of [key, value] pairs using dot/bracket notation.
- */
-function flattenEntries<T>(obj: T, path: string = ''): [string, unknown][] {
-  if (Array.isArray(obj)) {
-    return obj.flatMap((value, index) =>
-      flattenEntries(value, `${path}[${index}]`)
-    )
-  }
-
-  if (isObject(obj)) {
-    return Object.entries(obj).flatMap(([key, value]) =>
-      flattenEntries(value, path ? `${path}.${key}` : key)
-    )
-  }
-
-  return obj === undefined ? [] : [[path, obj]]
 }
 
 /**
